requirements.txt
# Database
psycopg2-binary==2.9.10
python-dotenv==1.0.0

# GUI
PyQt6==6.6.1

# Reporting
reportlab==4.0.8
openpyxl==3.1.2

# Data Generation
Faker==22.5.1

# Utils
bcrypt==4.1.2

qtawesome==1.3.1